+++ tensorly/__init__.py
@@ -2,7 +2,7 @@ import sys
 import importlib
 import os
 
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 
 # Set the default backend
 default_backend = 'numpy'
@@ -19,7 +19,7 @@ def set_backend(backend_name):
 
     Parameters
     ----------
-    backend_name : {'mxnet', 'numpy', 'pytorch'}, default is 'numpy'
+    backend_name : {'mxnet', 'numpy', 'pytorch', 'tensorflow', 'cupy'}, default is 'numpy'
     """
     global _BACKEND
     _BACKEND = backend_name
@@ -54,8 +54,5 @@ from .base import partial_tensor_to_vec, partial_vec_to_tensor
 
 from .kruskal_tensor import kruskal_to_tensor, kruskal_to_unfolded, kruskal_to_vec
 from .tucker_tensor import tucker_to_tensor, tucker_to_unfolded, tucker_to_vec
-
-
-