+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.4.0
+
+- Added TensorFlow backend with Eager execution
+- Added CuPy backend
+- PyTorch backend updated for PyTorch >= '0.4.0'
+
 # 0.3.0
 
 - Added PyTorch backend